% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/leMsmsRaw.R
\name{c.msmsWSspecs}
\alias{c.msmsWSspecs}
\title{Concatenation of raw spectra in msmsWorkspace}
\usage{
c.msmsWSspecs(w1, w2)
}
\arguments{
\item{w1,w2}{The msmsWorkspaces of which the spectra should be concatenated}
}
\value{
The \code{msmsWorkspace} with the spectra of two workspaces concatenated into one. Please note that the spectra from w2 will be concatenated to w1 and not vice versa.
}
\description{
Concatenates the (at)specs spectra of msmsWorkspace according to cpdIDs
}
\examples{
\dontrun{
		c.msmsWSspecs(w1,w2)
}
}
\author{
Erik Mueller
}
\seealso{
\code{\link{addPeaksManually}}
}

