% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundlist.R
\name{compoundlist2SDF}
\alias{compoundlist2SDF}
\title{Convert a Compoundlist into an SDF}
\usage{
compoundlist2SDF(filename)
}
\arguments{
\item{filename}{character
The name of the csv-file to be read. Note that the compoundlist
has to be filtered already.}
}
\value{
This method has no return value.
}
\description{
The resulting SDF will be written to a file named 'Compoundlist.sdf'.
The header for each block is the chemical name, tags for ID, SMILES and CAS
are added in the description block
}
\examples{
\dontrun{
			compoundlist2SDF("Compoundlist_filtered.csv")
}
}
\author{
pstahlhofen
}
