% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundlist.R
\name{createCompoundlist}
\alias{createCompoundlist}
\title{Create a Compoundlist from JCAMP files}
\usage{
createCompoundlist()
}
\value{
This method has no return value.
}
\description{
This method will automatically look for all single-block
JCAMP files in the directory by picking all files ending in '.dx'
(and not '.jdx'). A csv-file named 'Compoundlist.csv' will
be created in the same directory. The Compoundlist contains
columns 'ID', 'Name', 'SMILES' and 'CAS' where 'SMILES' might
be empty if the compound is a derivative or if the CAS number
could not be converted (see CAS2SMILES).
}
\examples{
\dontrun{
			# Prepare the compoundlist-creation
			splitMultiblockDX('my_multiblock_jcamp.jdx')
			createCompoundlist()
}
}
\seealso{
CAS2SMILES
}
\author{
pstahlhofen
}
