\name{dbe}
\alias{dbe}
\title{Calculate Double Bond Equivalents}
\usage{
  dbe(formula)
}
\arguments{
  \item{formula}{A molecular formula in text or list
  representation (e.g. \code{"C6H12O6"} or \code{list(C=6,
  H=12, O=6)} ).}
}
\value{
  Returns the DBE for the given formula.
}
\description{
  Calculates the Ring and Double Bond Equivalents for a
  chemical formula. The highest valence state of each atom
  is used, such that the returned DBE should never be below
  0.
}
\examples{
#
	dbe("C6H12O6")
}
\author{
  Michael Stravs
}

