% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmbSpectrum2Update.R
\name{.updateObject.RmbSpectrum2.formulaSource}
\alias{.updateObject.RmbSpectrum2.formulaSource}
\title{Add formulaSource column to spectrum.}
\usage{
.updateObject.RmbSpectrum2.formulaSource(w)
}
\arguments{
\item{w}{RmbSpectrum2
The object to be updated}
}
\value{
The updated RmbSpectrum2 object
}
\description{
TODO: consider whether to add functionality to move reanalysis stuff from legacy data back in.
}
\author{
stravsmi
}
