% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillback.R
\name{fillback}
\alias{fillback}
\title{Fill back reanalyzed / refiltered peak info into spectra}
\usage{
fillback(o, ...)
}
\description{
This method takes the info which is added to the aggregated table in the reanalysis and 
multiplicity filtering steps of the workflow, and adds it back into the spectra.
}
