% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundlist.R
\name{filterCompoundlist}
\alias{filterCompoundlist}
\title{Filter a Compoundlist for missing SMILES values}
\usage{
filterCompoundlist(filename)
}
\arguments{
\item{filename}{character
The name of the csv-file to be read}
}
\value{
This method has no return value.
}
\description{
Read the Compoundlist given by the filename and write a
'Compoundlist_filtered.csv', containing only the lines
with a SMILES string
}
\examples{
\dontrun{
			filterCompoundlist('Compoundlist.csv')
}
}
\author{
pstahlhofen
}
