% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/leMsmsRaw.R
\name{findMsMsHR.direct}
\alias{findMsMsHR.direct}
\title{Discontinued: find MS/MS spectrum from open raw file}
\usage{
findMsMsHR.direct(msRaw, cpdID, mode = "pH", confirmMode = 0,
  useRtLimit = TRUE,
  ppmFine = getOption("RMassBank")$findMsMsRawSettings$ppmFine,
  mzCoarse = getOption("RMassBank")$findMsMsRawSettings$mzCoarse,
  fillPrecursorScan = getOption("RMassBank")$findMsMsRawSettings$fillPrecursorScan,
  rtMargin = getOption("RMassBank")$rtMargin,
  deprofile = getOption("RMassBank")$deprofile, headerCache = NULL)
}
\arguments{
\item{msRaw}{x}

\item{cpdID}{x}

\item{mode}{x}

\item{confirmMode}{x}

\item{useRtLimit}{x}

\item{ppmFine}{x}

\item{mzCoarse}{x}

\item{fillPrecursorScan}{x}

\item{rtMargin}{x}

\item{deprofile}{x}

\item{headerCache}{x}
}
\value{
an error
}
\description{
This interface has been discontinued. \code{\link{findMsMsHR}} now supports the same parameters (use named
parameters).
}
\author{
stravsmi
}

