% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRecord.R
\name{getAnalyticalInfo}
\alias{getAnalyticalInfo}
\title{Get analytical info for MassBank record}
\usage{
getAnalyticalInfo(cpd = NULL)
}
\arguments{
\item{cpd}{A `RmbSpectraSet` object}
}
\description{
Collects the info for `ai, ac_lc, ac_ms` for general, LC and MS
 info respectively. The info comes from the settings except for the
 compound-specific part, which is omitted if there is no `cpd` specified.
}
