% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webAccess.R
\name{getCSID}
\alias{getCSID}
\title{Search ChemSpider CSID}
\usage{
getCSID(key, identifier, api_key)
}
\arguments{
\item{key}{ID to be converted}

\item{identifier}{identifier (name, inchikey)}

\item{api_key}{API key for ChemSpider (to be created on the developer site)}
}
\value{
The CSID (in string type)
}
\description{
Retrieves ChemSpider CSID from UK RSC for a search term.
}
\details{
Requires a valid API key
}
\examples{
\dontrun{
getDTXSID(key = "MKXZASYAUGDDCJ-NJAFHUGGSA-N", identifier = "InChIKey", api_key = "your RCS API key")
}
}
\references{
ChemSpider search: \url{https://developer.rsc.org/api-reference#}

ChemSpider developer site: \url{https://developer.rsc.org}

ChemSider REST:
\url{https://developer.rsc.org/api-reference#}
}
\author{
Tobias Schulze
}
