% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webAccess.R
\name{getDTXSID}
\alias{getDTXSID}
\title{Search CCTE DTXSID}
\usage{
getDTXSID(key, api_key)
}
\arguments{
\item{key}{ID to be converted}

\item{api_key}{API key for CCTE}
}
\value{
The DTXSID (in string type)
}
\description{
Retrieves CCTE DTXSID from US EPA for a search term.
}
\details{
Only the first result is returned currently. \bold{The function should be
regarded as experimental and has not thoroughly been tested.}
}
\examples{
\dontrun{
getDTXSID("MKXZASYAUGDDCJ-NJAFHUGGSA-N")
}
}
\references{
CCTE search: \url{https://api-ccte.epa.gov/docs}

CCTE REST:
\url{https://api-ccte.epa.gov/docs}
}
\author{
Tobias Schulze
}
