% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundlist.R
\name{getField}
\alias{getField}
\title{Get the content of a field in a JCAMP file}
\usage{
getField(parsedJDX, field_name)
}
\arguments{
\item{parsedJDX}{list as created by readJDX
A parsed, single-block JCAMP file}

\item{field_name}{character
The name of the field (e.g. 'CAS REGISTRY NO')}
}
\value{
The field's content
}
\description{
The content will always be returned as character-string
}
\examples{
\dontrun{
			parsedJDX <- readJDX('my_singleblock_jcamp.dx')
			title <- getField(parsedJDX, "TITLE")
}
}
\seealso{
readJDX
}
\author{
pstahlhofen
}
