% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_wrapper.R
\name{log_debug}
\alias{log_debug}
\title{Update logging file and pass arguments to logger::log_debug}
\usage{
log_debug(...)
}
\description{
The logging file to be used must be specified by the user in the \code{logging_file} field of \code{settings.ini}
}
\seealso{
logger::log_debug
}
\author{
pstahlhofen
}
