% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_wrapper.R
\name{log_info}
\alias{log_info}
\title{Update logging file and pass arguments to logger::log_info}
\usage{
log_info(...)
}
\description{
The logging file to be used must be specified by the user in the \code{logging_file} field of \code{settings.ini}
}
\seealso{
logger::log_info
}
\author{
pstahlhofen
}
