% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_wrapper.R
\name{log_warn}
\alias{log_warn}
\title{Update logging file and pass arguments to logger::log_warn}
\usage{
log_warn(...)
}
\description{
The logging file to be used must be specified by the user in the \code{logging_file} field of \code{settings.ini}
}
\seealso{
logger::log_warn
}
\author{
pstahlhofen
}
