% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMassBank.R
\name{makeMollist}
\alias{makeMollist}
\title{Write list.tsv file}
\usage{
makeMollist(compiled)
}
\arguments{
\item{compiled}{list of \code{RmbSpectraSet}
compiled spectra for multiple compounds (one \code{RmbSpectraSet} each).}
}
\value{
No return value.
}
\description{
Makes a list.tsv file in the "moldata" folder.
}
\details{
Generates the list.tsv file which is needed by MassBank to connect records with
their respective molfiles. The first compound name is linked to a mol-file with
the compound ID (e.g. 2334.mol for ID 2334).
}
\author{
Michael A. Stravs, Eawag <michael.stravs@eawag.ch>
}
