% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leMsmsRaw.R
\name{makePeaksCache}
\alias{makePeaksCache}
\title{Generate peaks cache}
\usage{
makePeaksCache(msRaw, headerCache)
}
\arguments{
\item{msRaw}{the input raw datafile (opened)}

\item{headerCache}{the cached header, or subset thereof for which peaks should be extracted. Peak extraction goes
by \code{seqNum}.}
}
\value{
A list of dataframes as from \code{mzR::peaks}.
}
\description{
Generates a peak cache table for use with \code{\link{findMsMsHR}} functions.
}
\author{
stravsmi
}
