% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSpectra.R
\name{mergePeaks.df}
\alias{mergePeaks.df}
\title{Merge peaks for spectra merging, FT shoulder elimination etc.}
\usage{
mergePeaks.df(peaks, dppm, dabs, int)
}
\description{
Note: ppm and abs are not cumulative!
}

