% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSpectra.R
\name{mergeSpectra}
\alias{mergeSpectra}
\alias{mergeSpectra,RmbSpectrum2List-method}
\title{Merge multiple spectra into one}
\usage{
mergeSpectra(spectra, ...)

\S4method{mergeSpectra}{RmbSpectrum2List}(spectra, ...)
}
\arguments{
\item{spectra}{\code{RmbSpectrum2List}
A list of \code{RmbSpectrum2} objects to be merged}

\item{...}{NOTHING
(This parameter is reserved for future implementations of the generic)}
}
\value{
A single \code{RmbSpectrum2} object
containing the merged information
}
\description{
This method takes a collection of \code{RmbSpectrum2} objects
and merges them into a single \code{RmbSpectrum2} object
}
\details{
Information from all spectra is retrieved via \code{\link{getData}}
combined with \code{rbind} and placed into the new spectrum with
\code{\link{setData}}
}
\seealso{
\code{\link{getData}}, \code{\link{setData}}
}
