% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\name{normalize,RmbSpectrum2-method}
\alias{normalize,RmbSpectrum2-method}
\title{Scale spectrum to specified intensity range}
\usage{
\S4method{normalize}{RmbSpectrum2}(object, ..., scale = 999, precision = 0, slot = "intensity")
}
\arguments{
\item{object}{the `RmbSpectrum2` object to scale}

\item{...}{arguments passed to `selectPeaks` to choose peaks for normalization}

\item{scale}{Maximum intensity in normalized spectrum}

\item{precision}{Digits after comma for normalized intensity, typically 0}

\item{slot}{Which property of the spectrum should be scaled}
}
\description{
Scale spectrum to specified intensity range
}
