% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\name{normalize,RmbSpectrum2List-method}
\alias{normalize,RmbSpectrum2List-method}
\title{Normalize spectra}
\usage{
\S4method{normalize}{RmbSpectrum2List}(object, ...)
}
\arguments{
\item{...}{Arguments passed to `normalize,RmbSpectrum2`}
}
\description{
Scale all spectra in a `RmbSpectrum2List` to a specified intensity.
}
