% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMassBank.R
\name{parseMassBank}
\alias{parseMassBank}
\title{MassBank-record Parser}
\usage{
parseMassBank(Files)
}
\arguments{
\item{Files}{array of character-strings
Paths to the plaintext-records that should be read}
}
\value{
The \code{mbWorkspace} that the plaintext-record creates.
All parsed information will be stored in the 'compiled_ok' slot.
}
\description{
Can parse MassBank-records(only V2)
}
\examples{
\dontrun{
	paths <- c("filepath_to_records/RC000001.txt",
	           "filepath_to_records/RC000002.txt")
		mb <- parseMassBank(paths)
}
}
\seealso{
\code{\link{validate}}
}
\author{
Erik Mueller
}
