% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMbRecord.R
\name{parseMbRecord}
\alias{parseMbRecord}
\title{MassBank-record Parser}
\usage{
parseMbRecord(filename, readAnnotation=TRUE)
}
\arguments{
\item{filename}{character
A path to the plaintext-record that should be read}

\item{readAnnotation}{logical, Default: TRUE
If TRUE, parse annotations from the record file and add columns for
'formula', 'formulaCount', 'mzCalc' and 'dppm' to the peak table}
}
\value{
An \code{RmbSpectrum2} object created from the plaintext-record
}
\description{
Can parse MassBank-records(only V2)
}
\examples{
\dontrun{
		parseMassBank("filepath_to_records/RC00001.txt")
}
}
\seealso{
\code{\link{validate}}
}
\author{
Erik Mueller
}
