% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\name{+,RmbSpectraSet,ANY-method}
\alias{+,RmbSpectraSet,ANY-method}
\title{Add a mass shift to a list of spectra}
\usage{
\S4method{+}{RmbSpectraSet,ANY}(e1, e2)
}
\arguments{
\item{e1}{a `RmbSpectraSet` object containing zero or more `children` spectra
and a `parent` spectrum}

\item{e2}{a numeric mass shift}
}
\description{
Shifts both `parent` and `children` spectra of the `RmbSpectraSet` by the 
same mass.
}
