% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\name{+,RmbSpectrum2List,ANY-method}
\alias{+,RmbSpectrum2List,ANY-method}
\title{Add a mass shift to a list of spectra}
\usage{
\S4method{+}{RmbSpectrum2List,ANY}(e1, e2)
}
\arguments{
\item{e1}{a `RmbSpectrum2List` object containing zero or more `RmbSpectrum2` spectra}

\item{e2}{a numeric mass shift}
}
\description{
Shifts all spectra in a `RmbSpectrum2List` by the same mass
}
