% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/SpectrumMethods.R
\name{property}
\alias{property}
\alias{property,RmbSpectrum2,character-method}
\title{Get a property of an RmbSpectrum2 object}
\usage{
property(o, property)

\S4method{property}{RmbSpectrum2,character}(o, property)
}
\arguments{
\item{o}{\code{RmbSpectrum2}}

\item{property}{character
The name of a property}
}
\value{
The corresponding column of \code{o@properties}
}
\description{
This searches the 'properties' slot of the object
and returns a column with matching name (if found)
or NULL otherwise.
}
