% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_wrapper.R
\name{rmb_log_info}
\alias{rmb_log_info}
\title{Pass arguments to logger::log_info using custom RMassBank-logging settings}
\usage{
rmb_log_info(...)
}
\description{
The logging file to be used can be specified by the user in the \code{logging_file} field of \code{settings.ini}
}
\seealso{
logger::log_info
}
\author{
pstahlhofen
}
