% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumClasses.R
\docType{class}
\name{RmbSpectraSet-class}
\alias{RmbSpectraSet-class}
\alias{.RmbSpectraSet}
\title{Set of spectra pertaining to one compound}
\description{
Set of spectra pertaining to one compound
}
\section{Slots}{

\describe{
\item{\code{parent}}{Spectrum1 The precursor spectrum}

\item{\code{children}}{RmbSpectrum2List List of `RmbSpectrum2` objects for the fragmentation
spectra, which are first extracted and later processed during `msmsWorkflow`}

\item{\code{found}}{logical, denotes whether or not fragmentation spectra were found for
this compound}

\item{\code{complete}}{logical, denotes whether or not *all* expected collision energies were
found for this compound}

\item{\code{empty}}{logical, `TRUE` if there are zero found spectra for this compound}

\item{\code{formula}}{character, the molecular formula of the neutral compound}

\item{\code{id}}{The ID of the compound in the RMassBank compound list (see \code{\link{loadList}})}

\item{\code{mz}}{the m/z value of the precursor}

\item{\code{name}}{The name of the compound}

\item{\code{mode}}{The ion type of the precursor, e.g. `pH, mH, mNa`}

\item{\code{smiles}}{the SMILES string for the compound structure}
}}

