% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumClasses.R
\docType{class}
\name{RmbSpectrum2List-class}
\alias{RmbSpectrum2List-class}
\alias{.RmbSpectrum2List}
\alias{RmbSpectrum2List}
\alias{RmbSpectraSetList}
\title{SimpleList specializations.}
\description{
Typed lists using `SimpleList`.
}
