% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\docType{methods}
\name{setData}
\alias{setData}
\alias{setData,RmbSpectrum2,data.frame-method}
\title{Set \code{RmbSpectrum2} data from data.frame}
\usage{
\S4method{setData}{RmbSpectrum2,data.frame}(s, df, clean = TRUE)
}
\arguments{
\item{s}{The \code{RmbSpectrum2} object to modify}

\item{df}{The data frame with new data}

\item{clean}{\code{TRUE} if slots which aren't present as columns in the data frame should be cleared.}
}
\value{
The modified \code{RmbSpectrum2}.
}
\description{
Sets all slots which are present as columns in the given dataframe. Optionally cleans the object, i.e. empties slots not defined in the data frame.
}
\author{
stravsmi
}

