% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumMethods.R
\name{-,Spectrum,numeric-method}
\alias{-,Spectrum,numeric-method}
\title{Add a negative mass shift to a spectrum}
\usage{
\S4method{-}{Spectrum,numeric}(e1, e2)
}
\arguments{
\item{e1}{a `MSnbase::Spectrum` object}

\item{e2}{a numeric mass shift}
}
\description{
Add a negative mass shift to a spectrum
}
