% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMassBank.R
\name{toMassbank,RmbSpectraSet-method}
\alias{toMassbank,RmbSpectraSet-method}
\title{Convert spectra from a `RmbSpectraSet` object to MassBank records}
\usage{
\S4method{toMassbank}{RmbSpectraSet}(o, addAnnotation = getOption("RMassBank")$add_annotation)
}
\arguments{
\item{addAnnotation}{`logical`, whether to add peak annotations (putative formulas) to the record.}
}
\value{
A list of character vectors with the MassBank records
}
\description{
Convert spectra from a `RmbSpectraSet` object to MassBank records
}
