% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMassBank.R
\name{toMassbank,RmbSpectrum2-method}
\alias{toMassbank,RmbSpectrum2-method}
\title{Convert a `RmbSpectrum2` object to a MassBank record}
\usage{
\S4method{toMassbank}{RmbSpectrum2}(o, addAnnotation = getOption("RMassBank")$add_annotation)
}
\arguments{
\item{addAnnotation}{`logical`, whether to add peak annotations (putative formulas) to the record.}
}
\value{
A character vector with the MassBank record
}
\description{
Convert a `RmbSpectrum2` object to a MassBank record
}
