% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundlist.R
\name{updateHeader}
\alias{updateHeader}
\title{Add a header to a Multiblock JCAMP file}
\usage{
updateHeader(filename)
}
\arguments{
\item{filename}{character
The name of the file to which a link block should be added.
The filename is also used as content for the TITLE field in the link block}
}
\value{
Nothing is returned
}
\description{
JCAMP files containing multiple blocks are usually structured
by so-called link blocks. If no link block is present, the readJDX
package is not able to parse the file. This method will add a link
block at the top of the given file or print a message if an existing
link block is found. The file is not changed in this case.
}
\examples{
\dontrun{
			updateHeader("my_multiblock_jcamp.jdx")
}
}
\author{
pstahlhofen
}
