% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/leMsMs.r
\name{cleanElnoise}
\alias{cleanElnoise}
\title{Remove electronic noise}
\usage{
cleanElnoise(peaks, noise=getOption("RMassBank")$electronicNoise,
		width = getOption("RMassBank")$electronicNoiseWidth)
}
\arguments{
\item{peaks}{An aggregated peak frame as described in \code{\link{aggregateSpectra}}. Columns
\code{mzFound}, \code{dppm} and \code{dppmBest} are needed.}

\item{noise}{A numeric vector of known m/z of electronic noise peaks from the instrument
Defaults to the    entries in the RMassBank settings.}

\item{width}{The window for the noise peak in m/z units. Defaults to the entries in
the RMassBank settings.}
}
\value{
Extends the aggregate data frame by column \code{noise} (logical), which is \code{TRUE} if the peak is marked as noise.
}
\description{
Removes known electronic noise peaks from a peak table
}
\examples{
# As used in the workflow:
\dontrun{
	    w@aggregated <-
		cleanElnoise(w@aggregated)
}
}
\author{
Michael Stravs
}
\seealso{
\code{\link{msmsWorkflow}}
}

