% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMassBank.R
\name{gatherCompound}
\alias{gatherCompound}
\alias{gatherSpectrum}
\title{Compose data block of MassBank record}
\usage{
gatherCompound(spec, aggregated, additionalPeaks = NULL)

		gatherSpectrum(spec, msmsdata, ac_ms, ac_lc, aggregated, 
 		additionalPeaks = NULL)
}
\arguments{
\item{spec}{A \code{RmbSpectraSet} object, representing a compound with multiple spectra.}

\item{aggregated}{An aggregate peak table where the peaks are extracted from.}

\item{additionalPeaks}{If present, a table with additional peaks to add into the spectra.
As loaded with \code{\link{addPeaks}}.}

\item{msmsdata}{A \code{RmbSpectrum2} object from the \code{spec} spectra set, representing a single spectrum to give a record.}

\item{ac_ms, ac_lc}{Information for the AC\$MASS_SPECTROMETRY and
AC\$CHROMATOGRAPHY fields in the MassBank record, created by
\code{gatherCompound} and then fed into \code{gatherSpectrum}.}
}
\value{
\code{gatherCompound} returns a list of tree-like MassBank data
blocks. \code{gatherSpectrum} returns one single MassBank data block or
\code{NA} if no useful peak is in the spectrum.
}
\description{
\code{gatherCompound} composes the data blocks (the "lower half") of all
MassBank records for a compound, using the annotation data in the RMassBank
options, spectrum info data from the \code{analyzedSpec}-type record and the
peaks from the reanalyzed, multiplicity-filtered peak table. It calls
\code{gatherSpectrum} for each child spectrum.
}
\details{
The returned data blocks are in format \code{list( "AC\$MASS_SPECTROMETRY" =
list('FRAGMENTATION_MODE' = 'CID', ...), ...)} etc.
}
\note{
Note that the global table \code{additionalPeaks} is also used as an
additional source of peaks.
}
\examples{
\dontrun{
     myspectrum <- w@spectra[[1]]
		massbankdata <- gatherCompound(myspectrum, w@aggregated)
		# Note: ac_lc and ac_ms are data blocks usually generated in gatherCompound and
		# passed on from there. The call below gives a relatively useless result :)
		ac_lc_dummy <- list()
		ac_ms_dummy <- list() 
		justOneSpectrum <- gatherSpectrum(myspectrum, myspectrum@child[[2]],
			ac_ms_dummy, ac_lc_dummy, w@aggregated)
}


}
\author{
Michael Stravs
}
\references{
MassBank record format:
\url{http://www.massbank.jp/manuals/MassBankRecord_en.pdf}
}
\seealso{
\code{\link{mbWorkflow}}, \code{\link{compileRecord}}
}

