% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webAccess.R
\name{getDTXCID}
\alias{getDTXCID}
\title{Search CCTE DTXCID}
\usage{
getDTXCID(key, api_key)
}
\arguments{
\item{key}{ID to be converted}

\item{api_key}{API key for CCTE}
}
\value{
The DTXCID (in string type)
}
\description{
Retrieves CCTE DTXCID from US EPA for a search term.
}
\details{
Only the first result is returned currently. \bold{The function should be
regarded as experimental and has not thoroughly been tested.}
}
\examples{
getDTXCID("MKXZASYAUGDDCJ-NJAFHUGGSA-N")

}
\references{
CCTE search: \url{https://api-ccte.epa.gov/docs} 

CCTE REST:
\url{https://api-ccte.epa.gov/docs}
}
\author{
Tobias Schulze
}
