% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webAccess.R
\name{getPcId}
\alias{getPcId}
\title{Search Pubchem CID}
\usage{
getPcId(query, from = "inchikey")
}
\arguments{
\item{query}{ID to be converted}

\item{from}{Type of input ID}
}
\value{
The PubChem CID (in string type).
}
\description{
Retrieves PubChem CIDs for a search term.
}
\details{
Only the first result is returned currently. \bold{The function should be
regarded as experimental and has not thoroughly been tested.}
}
\examples{
getPcId("MKXZASYAUGDDCJ-NJAFHUGGSA-N")

}
\references{
PubChem search: \url{http://pubchem.ncbi.nlm.nih.gov/} 

Pubchem REST:
\url{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}
}
\seealso{
\code{\link{getCtsRecord}}, \code{\link{getCactus}}
}
\author{
Michael Stravs, Erik Mueller
}
