\name{mbWorkflow}
\alias{mbWorkflow}
\title{MassBank record creation workflow}
\usage{
  mbWorkflow(mb, steps = c(1, 2, 3, 4, 5, 6, 7, 8),
    infolist_path = "./infolist.csv", useCTS = TRUE)
}
\arguments{
  \item{steps}{Which steps in the workflow to perform.}

  \item{infolist_path}{A path where to store newly
  downloaded compound informations, which should then be
  manually inspected.}

  \item{mb}{The \code{mbWorkspace} to work in.}
  
  \item{useCTS}{A boolean variable denoting whether 
  to retrieve information using CTS or to use babel	to
  retrieve minimal information.}

}
\value{
  The processed \code{mbWorkspace}.
}
\description{
  Uses data generated by \code{\link{msmsWorkflow}} to
  create MassBank records.
}
\details{
  See the vignette \code{vignette("RMassBank")} for
  detailed informations about the usage.

  Steps:

  Step 1: Find which compounds don't have annotation
  information yet. For these compounds, pull information
  from CTS (using gatherData).

  Step 2: If new compounds were found, then export the
  infolist.csv and stop the workflow.  Otherwise, continue.

  Step 3: Take the archive data (in table format) and
  reformat it to MassBank tree format.

  Step 4: Compile the spectra. Using the skeletons from the
  archive data, create MassBank records per compound and
  fill them with peak data for each spectrum.  Also, assign
  accession numbers based on scan mode and relative scan
  no.

  Step 5: Convert the internal tree-like representation of
  the MassBank data into flat-text string arrays
  (basically, into text-file style, but still in memory)

  Step 6: For all OK records, generate a corresponding
  molfile with the structure of the compound, based on the
  SMILES entry from the MassBank record. (This molfile is
  still in memory only, not yet a physical file)

  Step 7: If necessary, generate the appropriate
  subdirectories, and actually write the files to disk.

  Step 8: Create the list.tsv in the molfiles folder, which
  is required by MassBank to attribute substances to their
  corresponding structure molfiles.
}
\examples{
\dontrun{
		mb <- newMbWorkspace(w) # w being a msmsWorkspace
		mb <- loadInfolists(mb, "D:/myInfolistPath")
		mb <- mbWorkflow(mb, steps=c(1:3), "newinfos.csv")

}
}
\author{
  Michael A. Stravs, Eawag <michael.stravs@eawag.ch>
}
\seealso{
  \code{\link{mbWorkspace-class}}
}

