% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AggregateMethods.R
\docType{methods}
\name{peaksMatched}
\alias{peaksMatched}
\alias{peaksMatched,data.frame-method}
\alias{peaksMatched,msmsWorkspace-method}
\title{Select matching/unmatching peaks from aggregate table}
\usage{
peaksMatched(o)

\S4method{peaksMatched}{data.frame}(o)

\S4method{peaksMatched}{msmsWorkspace}(o)
}
\arguments{
\item{o}{Workspace or aggregate table from a workspace}
}
\value{
Selects the peaks from the aggregate table which matched within filter criteria (\code{peaksMatched}) or didn't match
		(\code{peaksUnmatched}).
}
\description{
Select matching/unmatching peaks from aggregate table
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{msmsWorkspace}: 
}}
\author{
stravsmi
}

