% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leMsMs.r
\name{problematicPeaks}
\alias{problematicPeaks}
\title{Identify intense peaks (in a list of unmatched peaks)}
\usage{
problematicPeaks(peaks_unmatched, peaks_matched, mode = "pH", retrieval="standard")
}
\arguments{
\item{peaks_unmatched}{Table of unmatched peaks, with at least \code{cpdID,
scan, mzFound, int}.}

\item{peaks_matched}{Table of matched peaks (used for base peak reference),
with at least \code{cpdID, scan, int}.}

\item{mode}{Processing mode (\code{"pH", "pNa"} etc.)}

\item{retrieval}{A value that determines whether the files should be handled either as "standard",
if the compoundlist is complete, "tentative", if at least a formula is present or "unknown"
if the only know thing is the m/z}
}
\value{
A filtered table with the potentially problematic peaks, including
the precursor mass and MSMS base peak intensity (\code{aMax}) for reference.
}
\description{
Finds a list of peaks in spectra with a high relative intensity (>10% and
1e4, or >1% and 1e5) to write a list of peaks which must be manually
checked.  Peaks orbiting around the parent peak mass (calculated from the
compound ID), which are very likely co-isolated substances, are ignored.
}
\examples{
\dontrun{
# As used in the workflow: 
fp <- problematicPeaks(specs[!specs$filterOK & !specs$noise & 
						((specs$dppm == specs$dppmBest) | (is.na(specs$dppmBest)))
				,,drop=FALSE], peaksMatched(w), mode)
}
}
\author{
Michael Stravs
}
\seealso{
\code{\link{msmsWorkflow}}
}

