% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leMsMs.r
\name{processProblematicPeaks}
\alias{processProblematicPeaks}
\title{Generate list of problematic peaks}
\usage{
processProblematicPeaks(w, mode, archivename = NA, retrieval = "standard")
}
\arguments{
\item{w}{\code{msmsWorkspace} to analyze.}

\item{mode}{Processing mode (pH etc)}

\item{archivename}{Base name of the archive to write to (for "abc" the exported failpeaks list will be "abc_Failpeaks.csv").}

\item{retrieval}{A value that determines whether the files should be handled either as "standard",
if the compoundlist is complete, "tentative", if at least a formula is present or "unknown"
if the only know thing is the m/z}
}
\value{
Returns the aggregate data.frame with added column "\code{problematic}" (logical) which marks peaks which match the problematic criteria
}
\description{
Generates a list of intense unmatched peaks for further review (the "failpeak list") and exports it if the archive name is given.
}
\author{
stravsmi
}

