% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leMsMs.r
\name{reanalyzeFailpeaks}
\alias{reanalyzeFailpeaks}
\alias{reanalyzeFailpeak}
\title{Reanalyze unmatched peaks}
\usage{
reanalyzeFailpeaks(aggregated, custom_additions, mode, filterSettings =
				getOption("RMassBank")$filterSettings, progressbar = "progressBarHook")
reanalyzeFailpeak(custom_additions, mass, cpdID, counter, pb = NULL, mode,
				filterSettings = getOption("RMassBank")$filterSettings)
}
\arguments{
\item{custom_additions}{The allowed additions, e.g. "N2O".}

\item{mode}{Processing mode (\code{"pH", "pNa", "mH"} etc.)}

\item{filterSettings}{Settings for filtering data. Refer to\code{\link{analyzeMsMs}} for settings.}

\item{progressbar}{The progress bar callback to use. Only needed for specialized
applications.    Cf. the documentation of \code{\link{progressBarHook}} for usage.}

\item{aggregated}{A peake aggregate table (\code{w@aggregate}) (after processing electronic noise removal!)}

\item{mass}{(Usually recalibrated) m/z value of the peak.}

\item{cpdID}{Compound ID of this spectrum.}

\item{counter}{Current peak index (used exclusively for the progress
indicator)}

\item{pb}{A progressbar object to display progress on, as passed by
 \code{reanalyzeFailpeaks} to \code{reanalyzeFailpeak}. No progress 
is displayed if NULL.}
}
\value{
The aggregate data frame extended by the columns:
#' \item{reanalyzed.???}{If reanalysis (step 7) has already been processed: matching values from the reanalyzed peaks}
\item{matchedReanalysis}{Whether reanalysis has matched (\code{TRUE}), not matched(\code{FALSE}) or has not been conducted for the peak(\code{NA}).}

It would be good to merge the analysis functions of \code{analyzeMsMs} with
the one used here, to simplify code changes.
}
\description{
Reanalysis of peaks with no matching molecular formula by allowing
additional elements (e.g. "N2O").
}
\details{
\code{reanalyzeFailpeaks} examines the \code{unmatchedPeaksC} table in
\code{specs} and sends every peak through \code{reanalyzeFailpeak}.
}
\examples{

## As used in the workflow:
\dontrun{    
	reanalyzedRcSpecs <- reanalyzeFailpeaks(w@aggregated, custom_additions="N2O", mode="pH")
# A single peak:
reanalyzeFailpeak("N2O", 105.0447, 1234, 1, 1, "pH")
}







}
\seealso{
\code{\link{analyzeMsMs}}, \code{\link{msmsWorkflow}}
}
\author{
Michael Stravs
}
