% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateMethods.R
\docType{methods}
\name{selectPeaks}
\alias{selectPeaks}
\alias{selectPeaks,data.frame-method}
\alias{selectPeaks,msmsWorkspace-method}
\title{Select peaks from aggregate table}
\usage{
selectPeaks(o, ...)

\S4method{selectPeaks}{data.frame}(o, good = FALSE, bad = FALSE,
  cleaned = FALSE, best = FALSE)

\S4method{selectPeaks}{msmsWorkspace}(o, ...)
}
\arguments{
\item{o}{\code{msmsWorkspace} or aggregate \code{data.frame} from a workspace.}

\item{...}{no additional parameters}

\item{good}{if \code{TRUE}, include good (matched within filter criteria) peaks.}

\item{bad}{if \code{TRUE}, include bad (not matched within filter criteria) peaks. Note: \code{good} and \code{bad} can be
combined, both are returned in that case.}

\item{cleaned}{if \code{TRUE}, return only peaks which passed the noise filter. Note: If the noise filter was not applied, the
parameter has no effect. Also, a \code{noise} column is in any case added to the output, even if not present before.}

\item{best}{if \code{TRUE}, only select the best match for each peak (i.e. the formula with smallest delta ppm). Otherwise multiple
matches can be returned.}
}
\value{
Peak dataframe according to the specified criteria.
}
\description{
Selects peaks from aggregate table according to different criteria.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: A method to retrieve the specified peaks from the "aggregated" slot (a data.frame object) in an msmsWorkSpace

\item \code{msmsWorkspace}: A method to retrieve the specified peaks from an msmsWorkSpace
}}
\author{
stravsmi
}

